Some things:
(1) Tidied the CIF up, with spaces between groups of related items and that sort of thing.

(2) Call XSFLSB at start to update L30 calc-R,Rw,sigma and number.

(3) Calculate ZPRIME from L2 and L30 Z value, if L30 Z value is non-zero. Use this
to give desired formula.

(4) Quote three groups of R-factors: _ref, _all and _gt. The _gt threshold is the
same as the L28 threshold if present, otherwise it is I>4u(I).


git-svn-id: https://svn.debroglie.net/crystals/trunk@1687 4ce1adb0-6e93-48c6-b531-a91e7d66e1f5