In #DISTANCE added option: SEL RANGE=L41 This will use the bond list (41) to determine what is connected.
In #DISTANCE added option: OUT PUNCH=DELU D1DEV=x D2DEV=y
This will output VIBR restraints to the punch file with an esd
of x for 1,2 distances and y for 1,3 distances. You probably want an
EXCL H if using the DELU option.

In #DISTANCE added option: OUT PUNCH=SIMU S1DEV=x S2DEV=y
This will output U(IJ) restraints to the punch file with an esd
of x for normal bonds and y for terminal bonds. A bond is terminal
if it connects to an atom with no other connections except hydrogen.
This option uses LIST 41 to determine whether the atom is 'terminal',
but the RANGE can still be set at user's discretion. You probably want an
EXCL H if using the SIMU option.

In #EDIT added option: INSERT NCONN
This will put the number of connections to other atoms (according to L41)
into the SPARE slot in List 5.

In #EDIT added option: INSERT RELAX
This will start with the electron count of each atom in SPARE, and
keep cycling through the bond list, adding the current value of
SPARE to all the neighbouring atoms. It will cycle until the number
of unique values of SPARE stops increasing.

distangl.src now contains a routine KDIST4, which works like KDIST1 but using
the existing L41. It should be noted that the bond lengths stored in L41
may be out of date as the bonding is only recalculated every now and again.


git-svn-id: https://svn.debroglie.net/crystals/trunk@1731 4ce1adb0-6e93-48c6-b531-a91e7d66e1f5