To List 12 and List 16, the directive 'REM' has been added. Anything following this directive is ignored, but it still stored in list 12 or 16 respectively.
To List 12, the directive 'SUMFIX' has been added. SUMFIX fixes the sum of
groups of parameters, so that it doesn't change from it's current value.
E.g. SUMFIX O(1,OCC) H(1,OCC) AND O(2,OCC) H(2,OCC) AND O(3,OCC) H(3,OCC)
where three hydroxyl groups are disordered across three sites. The groups
are separated by the keyword AND. The user must ensure that occupancies
are set correctly in advance.
As a special bonus, if AND is omitted all together, then the code imagines
that you forgot to type it in between every parameter on the directive. E.g.
SUMFIX O(1,OCC) O(2,OCC) O(3,OCC) fixes the sum of those three parameters and
is equivalent to SUMFIX O(1,OCC) AND O(2,OCC) AND O(3,OCC). Although this reduces
the ability of the directive to apply to groups of parameters, it makes
some restraints very simple:
E.g. SUMFIX ELEMENT SCALES fixes the sum of the twin element scales.
E.g. SUMFIX ALL(X) fixes the sum of the x-coordinates.


git-svn-id: https://svn.debroglie.net/crystals/trunk@2226 4ce1adb0-6e93-48c6-b531-a91e7d66e1f5