Distances and angles in CIF now use List 41.  As a results PART information is used to decide which atoms are bonded.
One problem to be aware of: An old version of the GUIDE used to
erroneously set a value in the 'PART' slot of list 5 when
setting up a refinement. The symptom this causes now is that
some obvious bonds are missing (e.g. quite often C-H). The
following will fix it: #EDIT/CHANGE PART 0/END


git-svn-id: https://svn.debroglie.net/crystals/trunk@2237 4ce1adb0-6e93-48c6-b531-a91e7d66e1f5