Added a 'ROTATE' directive to #EDIT. There are three modes of operation:
  rotate angle px py pz vx vy vz [atoms]
  rotate angle patom vx vy vz [atoms]
  rotate angle patom vatom [atoms]

The first rotates the specified [atoms], angle degrees around the vector
vx,vy,vz keeping point px,py,pz fixed. (px-pz and vx-vz are given in
crystal fractions).

The rotation is in an orthogonal system so all distances and angles
within the moved set of atoms are preserved.

The second notation replaces the point to keep fixed with patom, the name
of an atom. Its coordinates are used to obtain px py and pz.

The third notation uses patom to get px, py and pz, and the vector from patom
to vatom to get vx, vy and vz.

Examples:

1) Rotate the hydrogens of a methyl group by sixty degrees.

    #EDIT
    ROTATE 60 C(1) C(2) H(20) H(21) H(22)
    END

2) Turn a phenyl ring through 30 degrees around its rotation axis.

    #EDIT
    ROTATE 30 C(1) C(20) C(21) C(22) C(23) C(24) C(25)
    END

3) Flip a residue 90 degrees around the a-axis about its centroid, QC(1)
   (see also CENTROID and INSERT RESIDUE directives)

    #EDIT
    INSERT RESIDUE
    CENTROID 1 RESIDUE(1)
    ROTATE 90 QC(1) 1 0 0 RESIDUE(1)
    END


git-svn-id: https://svn.debroglie.net/crystals/trunk@2723 4ce1adb0-6e93-48c6-b531-a91e7d66e1f5