Enable use of dual wavelength data
Dual Wavelength Refinement.		Sept 1010

Dual wavelength machine are becoming increasingly common.  This provides
the possibility of simultaneously carrying out a high resolution refinement
(with the Mo data) with good leverage on the hydrogen atoms and the absolute
configuration (Cu data).

Data Storage:

Reflections for both Mo and Cu data.
Extinction correction for each radiation
Batch scale factor for each radiation
Weighting scheme for each radiation
f' and f" for each radiation

Programming issues:

Merge:  Ability to merge reflections keeping the different wavelengths separate.
Agreement Analysis on each wavelength separately
Fourier.  All data for one or other wavelength.
Fourier.  Possibly 'correct' Fo for extinction and then merge scaled data
	for both radiations.
Ton Speks Bivjoet code.  Handle wavelengths separately or sequentially.



Problems with CRYSTALS implementation:

1. No proper place to store f' f" for secondary radiation.
2. No mechanism for storing value of secondary wavelength
3. No mechanism for indicating wavelength used for each reflection
4. Only one extinction correction stored

1. f' f" currently tagged on to the end of a LIST 3 record.
2. Assume Mo is principal and Cu secondary wavelngth
3. Use the 'BATCH'' identifier as the wavelength selector (1=Mo, 2=Cu)
and the batch scalefactors as the wavelength scalefaactors
4. There are unused OVERALL parameters for anisotropic extinction (for
neutron work).  This was never completed because there was (at that time)
no general agreement about the best expression for extinction.

Insuperable(?) CRYSTALS Problem:

	CRYSTALS puts each new reflection into a stack as its derivatives
are computed, possibley as a mechanism to save recomputing bits when
twinned data is used.  Each new reflection is checked against the stack,
and if its indices already occur, derivatives etc are not calculated.
For a data set with data from Mo and Cu radiation, a reflecion is added
to the stack for the first one encountered, and then the precomputed
values used for the second one.  In effect the second wavelength is not used.

Redesign Features

For each reflection h,k,l store lambda and  the intensity and sigma
for the reflection itself and its Friedel pair for each wavelength.
This would lead to significant economy at execution time because most
of the expression is wavelength independent.


git-svn-id: https://svn.debroglie.net/crystals/trunk@4214 4ce1adb0-6e93-48c6-b531-a91e7d66e1f5