Revert convplat.F to not use .fab files.
Code was previously modified to read platon_sq.fab files in preference to platon_sqd.hkl output. However, the .fab indexing and ordering can be signficantly different to the hkl, and that isn't handled in this code. Therefore reverting (Jan19) to reading the platon_sqd.hkl file which contains a and b parts beyond col 80. Apr18 code is just commented out for now.