H finding improvements (including neutron)
GFINDH improvements: Added facility to model to display peaks based on max cutoff (for negative neutron scatterers).
Added facility to CrModel to hide peaks that are not bonded to anything, or are only bonded to other peaks. Used in gfindh to clear the view of the maps.
Added Q to gfindh for neutron structures - are you looking for D or H. Use NMAP as required. Added slider and code to allow negative thresholds. Added control to hide lone peaks. Added control to hide the info text at the top of the page to increase size of structure view.